+++ demisto_sdk/commands/common/tests/playbook_test.py
@@ -310,6 +310,28 @@ class TestPlaybookValidator:
         (DEPRECATED_INVALID_DESC3, False)
     ]
 
+    CONDITIONAL_SCRPT_WITH_DFLT_NXT_TASK = {"id": "Intezer - scan host", "version": -1,
+                                            "tasks":
+                                                {'1': {'type': 'condition',
+                                                       'scriptName': 'testScript',
+                                                       'nexttasks': {'#default#': []}}}}
+
+    CONDITIONAL_SCRPT_WITH_NO_DFLT_NXT_TASK = {"id": "Intezer - scan host", "version": -1,
+                                               "tasks":
+                                               {'1': {'type': 'condition',
+                                                      'scriptName': 'testScript',
+                                                      'nexttasks': {'1': []}}}}
+
+    CONDITION_TASK_WITH_ELSE = {'1': {'type': 'condition',
+                                      'scriptName': 'testScript',
+                                      'nexttasks': {'#default#': []}}}
+
+    CONDITION_TASK_WITHOUT_ELSE = {'1': {'type': 'condition',
+                                         'scriptName': 'testScript',
+                                                       'nexttasks': {'1': []}}}
+    IS_ELSE_IN_CONDITION_TASK = [(CONDITIONAL_SCRPT_WITH_NO_DFLT_NXT_TASK.get('tasks').get('1'), False),
+                                 (CONDITIONAL_SCRPT_WITH_DFLT_NXT_TASK.get('tasks').get('1'), True)]
+
     @pytest.mark.parametrize("playbook_json, id_set_json, expected_result", IS_SCRIPT_ID_VALID)
     def test_playbook_script_id(self, mocker, playbook, repo, playbook_json, id_set_json, expected_result):
         """
@@ -463,3 +485,36 @@ class TestPlaybookValidator:
         validator = PlaybookValidator(structure)
         validator.current_file = current
         assert validator.is_valid_as_deprecated() is answer
+
+    @pytest.mark.parametrize("playbook_json, expected_result", [(CONDITIONAL_SCRPT_WITH_NO_DFLT_NXT_TASK, True)])
+    def test_verify_all_conditional_tasks_has_else_path(self, playbook_json, expected_result):
+        """
+        Given
+            - A playbook with a condition without a default task
+
+        When
+            - Running Validate playbook
+
+        Then
+            - Function returns true as this is an ignored error.
+        """
+        structure = mock_structure("", playbook_json)
+        validator = PlaybookValidator(structure)
+        assert validator.verify_condition_tasks_has_else_path() is expected_result
+
+    @pytest.mark.parametrize("playbook_task_json, expected_result", IS_ELSE_IN_CONDITION_TASK)
+    def test_verify_else_for_conditions_task(self, playbook_task_json, expected_result):
+        """
+        Given
+            - A playbook condition task with a default task
+            - A playbook condition task without a default task
+
+        When
+            - Running Validate playbook
+
+        Then
+            - Return True if the condition task has default path , else false
+        """
+        structure = mock_structure("", playbook_task_json)
+        validator = PlaybookValidator(structure)
+        assert validator._is_else_path_in_condition_task(task=playbook_task_json) is expected_result